Bigbig Refactor to CustomHooks and db optimised
