--- dictie/db.tsx
@@ -1,126 +1,0 @@
-export const givenText = {
-  easy: {
-    boolValue: false,
-    arrayValue: [
-      "șase sași în șase saci",
-      "O babă bălană mănâncă o banană babană.",
-      "Cupă cu capac, capac cu cupă.",
-      "Nașpa, nasol stau în nămol ca un actor jucând un rol!",
-      "Clopotarul clătina clopotul clopotniței.",
-    ],
-  },
-  easyOne: {
-    boolValue: false,
-    arrayValue: [
-      "Retevei de tei pe mirişte de mei.",
-      "Zecele de piscă mare are masa la picioare.",
-      "O barză brează face zarvă pe-o varză",
-      "Bucur și Bucura se bucură că Bucurel e bucuros în București.",
-      "Leul lăudat lipește abțibilduri lucioase.",
-    ],
-  },
-  easyTwo: {
-    boolValue: false,
-    arrayValue: [
-      "Cinci cocosi cocosati cocoseste cinci cocose.",
-      "Un vultur stă pe-un pisc c-un pix în plisc",
-      "Domnule dudar, dă-mi două dude dulci din dudul dumitale de dincolo de Dunăre!",
-      "Capra calca-n piatra,Piatra crapa-n patru; Crăpa-i-ar capul caprei, Precum a crapat piatra-n patru.",
-      "Unui tâmplar i s-a întâmplat o întâmplare. \nAlt tâmplar, auzind de întâmplarea tâmplarului \nde la tâmplărie, a venit și s-a lovit cu tâmpla de \n tâmplăria tâmplarului cu întâmplarea",
-    ],
-  },
-  easyThree: {
-    boolValue: false,
-    arrayValue: [
-      "Prin Vulturi vântul viu vuia",
-      "Leul lăudat lipește abțibilduri lucioase.",
-      "Sasha şuşotea șosetelor însuşiri şiruitoare.",
-      "Am o prepeliţă pestriţă cu paisprezece pui de prepeliţă pestriţi. E mai pestriţă prepeliţa pestriţă decât cei paisprezece pui de prepeliţă pestriţi.",
-      "O coropișniță și-un coropișnițoi se coropișniteau la noi pe gunoi. Nu coropișnița coropișnițea pe coropișnițoi, ci coropișnițoiul coropișnițea pe coropișniță.",
-    ],
-  },
-  easyFour: {
-    boolValue: false,
-    arrayValue: [
-      "Greierii grăiască în grai greieresc.",
-      "La șase șosele s-adunau șase sute de șoșoni.",
-      "Nu știu dacă știi să știi, cu a ști se știe că știu.",
-      "Preoteasa Popii din Popeasa prăjea prăjituri pe prispă.",
-      "Treizeci și trei de trepiede treieră treizeci și trei de trepte",
-      "Vremea trecea, timpul venea, dacă vremea n-ar trece, timpul n-ar veni.",
-    ],
-  },
-  easyFive: {
-    boolValue: false,
-    arrayValue: [
-      "Cinci fine meduze  / Geometrii confuze / Cu unghiuri obtuze / Și ipotenuze",
-      "Șase călăuze / Încearcã s-acuze / Cinci femei lehuze / Că ar fi farfuze",
-      "Cele șapte muze / Cu priviri de zuze / Scriu ca să se amuze / Versuri andaluze",
-      "Patru buburuze / Roşii archebuze / Bâzâie ursuze / Zburând pe peluze",
-      "Trei babe mofluze / Fără dinți și buze / Trag turte pe spuze / Fără să se scuze",
-      "Douã autobuze / Lovite de-obuze / Au lumini difuze / Și defect la diuze",
-      "Una din ecluze / La uși cu mezuze / Lupta cu taluze / Punându-și ventuze",
-    ],
-  },
-  easySix: {
-    boolValue: false,
-    arrayValue: [
-      "O babă are o batistă boțită.",
-      "Cinci cocosi cocosati cocoseste cinci cocose.",
-      "Corcodușele clocotind se coc în clocote clocotitoare.",
-      "Greu de tot e să tot zici greu de tot, și totuși e tot mai greu să tot zici tot mai greu.",
-      "S-a suit Șoarecele Susu pe un susan sus și i-a șoptit Șoaricii Sese să suspende surubul în sus.",
-      "Un bal fara egal cu final fatal la un halal carnaval estival cu scandal epocal dintr-un opal oval, pal, real si natural, fara rival, egal si actual.",
-      "Rica nu stia si zica: rau, ratusca, ramurica, dar de cand baiatu-nvata poezia despre rata, Rica stie acum sa zica: rau, ratusca, ramurica."
-    ],
-  },
-  easySeven: {
-    boolValue: false,
-    arrayValue: [
-      "Am aflat că Ana are un ananas.",
-      "Gardul gârbovit gâdilă gâgâind gâgâitorul",
-      "Trenul trăgând tren tendențios trecea de-atâtea ori.",
-      "Ghita, caprita, mananca iedera pana nu o sa mai fie verde.",
-      "Vioara varia în zadar variațiuni variate vara în văgăuna văratecului.",
-      "Vin vânzătorii de vată în vânt cu vara, vin și vânzătorii de vată în vânt cu vișine?",
-      "Într-un sat fără câini, cocostârcii venind de la câmp, nu găsesc pișcot să se așeze ci să picotească.",
-    " Șusotind scenariul scenetei fosforescente, Sasha suferea de un exces de efervescență și de sârlul sacalului suparat."
-    ],
-  },
-  easyEight: {
-    boolValue: false,
-    arrayValue: [
-      "Buburuza Zuza zâmbind zumzanea",
-      "Văduva villei va vinde vila vilei.",
-      "Zevzecul cumpără zarzavaturile zarzavagiului.",
-      "Zece zidari zidesc, zidind zdravăn zidul zicând: 'Zidul zdraven! Zidul zidit!'",
-      "Pe un pod de fier peste un pârâu de piatră trece-un podar cu un porc într-un picior de paltin.",
-      "Papucarul papucareste papucii papucaresei, papucareasa nu poate papucari papucii papucarului.",
-      "Cupa cu pupăza, pupăza în cupă, și-a rupt o aripă și nu mai pupă.",
-      "Sergiu cel sănătos și musculos sare peste stâlp, și-n sus și-n jos, săsând sistematic un susur scos din senin, suspicios și sfios",
-      "Crusta crustaceelor creșsează descrescător un piscicultor."
-    ],
-  },
-  easyNine: {
-    boolValue: false,
-    arrayValue: [
-      "Curcanul curcă curcând curca curcanului.",
-      "Chipul cioplit în chiparos chiparositos.",
-      "În Grădina Gheții gângăveau greierii gheții",
-      "Șapte șopârle săpate sapă săpături săptămânale.",
-      "Fără frică, în frig frământă frământată fără frișcă.",
-      "Fă frumușică firimitură fără să frângi frământătura.",
-     
-      "Arcașul țintar își întinde arcuşul să întindă întinsul.",
-      "Rebătător, răzbunător, arzător, revoltat, arici iritat.",
-      "Pot să fac o sută de feluri de mâncare fără să flămânzesc.",
-      "Exercițiul exersat excede incidența exercitării exercițiului.",
-      "Un vânător știa că vânatul șvâncit în vânatărie nu va veni niciodată să vâneze.",
-      "Care crâncen crâncenează în crânceni crâncenul ce crâncenează crâncenii crâncenind crâncen",
-      "Titularizarea titlului deteriorat, tipărit total, traumatizează tot turnul căzut pe sat."
-
-
-
-    ],
-  }
-};
