+++ dictie/src/App.tsx
@@ -1,22 +1,17 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import LandingPage from "./pages/LandingPage";
-import Exercise from "./components/Exercise";
+import Exercise from "./components/exercise/Exercise";
 import ExercisesPage from "./pages/ExercisesPage";
-import { CollectionContext  } from "./context/CollectionContext";
-import { useState } from "react";
+import { CollectionProvider } from "./provider/CollectionProvider";
 
-type CollectionContextType = {
-  document: string;
-  setDocument: (doc: string) => void;
-};
 
 
 function App() {
 
-  const [document, setDocument] = useState("exercises-easy");
+
 
   return (
-    <CollectionContext.Provider value={{ document, setDocument } as CollectionContextType}>
+  <CollectionProvider>
       <Router>
         <Routes>
           <Route path="/" element={<LandingPage />} />
@@ -24,8 +19,8 @@ function App() {
           <Route path="/exercises" element={<ExercisesPage />} />
         </Routes>
       </Router>
-    </CollectionContext.Provider>
-  );
+      </CollectionProvider>
+  )
 }
 
 export default App;
