--- dictie/src/components/GreuComponent.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-// import { Link } from 'react-router-dom';
-
-const GreuComp: React.FC = () => {
-    return (
-        <div>
-            <h1>-Coming soon-</h1>
-            <ul>
-                {/* <Link to={'/easy'}><li>sase sasi</li></Link>
-                <li>unu doi greu</li>
-                <li>pisica pisiceste greoi</li>
-                <li>castelul castelar greu</li>
-                <li>capra greu</li>
-                <li>caprarur greu</li> */}
-            </ul>
-        </div>
-    );
-};
-
-export default GreuComp;
