+++ dictie/src/pages/ExercisesPage.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import Header from "../components/Header";
 import "../style/style.exercise.css";
-import ExerciseList from "../components/ExerciseList";
+import ExerciseList from "../components/exercise/ExerciseList";
 
-import GreuComp from "../components/GreuComponent";
 import { useContext } from "react";
 import { AuthContext } from "../context/AuthContext";
 import { CollectionContext } from "../context/CollectionContext";
@@ -13,11 +12,10 @@ type CollectionContextType = {
   setDocument: (doc: string) => void;
 };
 const ExercisesPage: React.FC = () => {
-
   const user = useContext(AuthContext);
-  // ?????????????? how tf to fix this type error
-  const { document, setDocument } = useContext<CollectionContextType>(CollectionContext as unknown as React.Context<CollectionContextType>);
- 
+  const { document, setDocument } = useContext<CollectionContextType>(
+    CollectionContext as unknown as React.Context<CollectionContextType>
+  );
 
   return (
     <div>
@@ -32,18 +30,11 @@ const ExercisesPage: React.FC = () => {
           </ul>
         </div>
         <div className="right-main">
-          {document === "exercises-easy" && <ExerciseList title="Usor"  />}
+          {document === "exercises-easy" && <ExerciseList title="Usor" />}
 
-
-          {user ? (
-           document == "exercises-mid" && <ExerciseList title="Mediu"  />
-          ) : (
-           document == "exercises-mid" && <h1>Log in to unlock</h1>
-          )}
-         
-         
-         
-          { document == "exercises-hard" && <GreuComp />}
+          {user
+            ? document == "exercises-mid" && <ExerciseList title="Mediu" />
+            : document == "exercises-mid" && <h1>Log in to unlock</h1>}
         </div>
       </div>
     </div>
