+++ dictie/src/pages/VoiceInput.tsx
@@ -1,62 +1,40 @@
 import React, { useCallback, useEffect, useRef, useState } from "react";
 import handleVoiceInput from "../utils/handleVoice.tsx";
 import addAnimation from "../utils/addAnimation.tsx";
-
+import { useSetFeeling } from "../hooks/useSetFeeling.tsx";
 import "../style/style.animations.css";
 import "../style/App.css";
 import Header from "../components/Header.tsx";
 import Animations from "../components/Animations.tsx";
-import { Link  } from "react-router-dom";
+import { Link } from "react-router-dom";
 import percentageFunc from "../utils/percentCalc.tsx";
 import manageScore from "../utils/manageScore.tsx";
-
+import ScoreBoard from "../components/exercise/ScoreBoard.tsx";
+import { useScoreBoard } from "../hooks/useScoreBoard.tsx";
+import ExerciseFinish from "../components/exercise/ExerciseFinish.tsx";
+import ExerciseQuestions from "../components/exercise/ExerciseQuestions.tsx";
 
 interface VoiceInputProps {
   questions: string[];
 }
 
-
 const VoiceInput: React.FC<VoiceInputProps> = ({ questions }) => {
-  //  REFACTOR THIS COMPONENT AGAIN
-
-
-
-  
-
-
-
-  console.log("FUNCTION START!!!!!!!!!!!!!");
-
   // State from handle voice
   const [spokenText, setSpokenText] = useState<string>("");
   const [speechEnd, setSpeechEnd] = useState<boolean>(true);
-
-  console.log(speechEnd, "Speech End");
-  console.log(spokenText, "Spoken Text");
-
   const [animationKey, setAnimationKey] = useState<number>(0);
-  console.log(animationKey, "Animation Key");
-
-  const [userScore, setUserScore] = useState<number>(0);
 
   const index = useRef<number>(0);
-  console.log(index.current, "Current Index");
-
   const [feeling, setFeeting] = useState<string>("😊 Sa incepem");
-  console.log(feeling, "Feeling");
-
   const similar = percentageFunc(spokenText, questions[index.current]);
-  console.log(similar, "Similarity");
-
+  const useSetFeelingValue = useSetFeeling(similar, feeling, setFeeting);
   const scoreboard = manageScore(similar);
-  console.log(scoreboard);
+  const userScore = useScoreBoard({ scoreboard, speechEnd });
 
-  // Change this to context in future
   const resetAnimation = useCallback(() => {
     // Incrementing the animationKey will re-render the component
     setAnimationKey((prevKey) => prevKey + 1);
   }, [setAnimationKey]);
-
 
   const resetValues = () => {
     resetAnimation();
@@ -79,6 +57,7 @@ const VoiceInput: React.FC<VoiceInputProps> = ({ questions }) => {
     index.current++;
   }, [resetAnimation]);
 
+  // Wait half a sec for the speech to end and then go to the next question
   useEffect(() => {
     setTimeout(() => {
       if (speechEnd && spokenText) {
@@ -87,33 +66,6 @@ const VoiceInput: React.FC<VoiceInputProps> = ({ questions }) => {
       }
     }, 500);
   }, [speechEnd]);
-
-  // Change text based on similarity
-  useEffect(() => {
-    if (similar === 0) return;
-    similar < 5
-      ? setFeeting("😊 Sa incepem")
-      : similar <= 50
-      ? setFeeting("😔 Poți face mai bine")
-      : similar > 50 && similar < 80
-      ? setFeeting("😯 Te apropii!")
-      : similar > 80 && similar < 90
-      ? setFeeting("😃 Lucru excelent!")
-      : similar > 90
-      ? setFeeting("😍 Ai făcut uimitor!")
-      : setFeeting("");
-  }, [similar]);
-
-  useEffect(() => {
-    if (scoreboard === "point" && speechEnd) {
-      const scoreboard = document.querySelector(".score-board") as HTMLElement;
-      scoreboard.style.color = "green";
-      setUserScore((prev) => prev + 1);
-    } else if (scoreboard === "no point" && speechEnd) {
-      const scoreboard = document.querySelector(".score-board") as HTMLElement;
-      scoreboard.style.color = "red";
-    }
-  }, [speechEnd, scoreboard]);
 
   const backBtn = (
     <Link to="/exercises">
@@ -129,93 +81,84 @@ const VoiceInput: React.FC<VoiceInputProps> = ({ questions }) => {
     });
   };
 
-  console.log(similar);
-
   return (
     <>
       <Animations animationKey={animationKey} />
       <Header backButton={backBtn} />
       <main>
-        <span className="score-board">
-          {userScore}/{questions.length}
-        </span>
-        {/* IF THE EXERcIS IS OVER DISPLAY END */}
+        <ScoreBoard questions={questions} userScore={userScore} />
+        {/* IF THE EXERCISE IS OVER = DISPLAY END */}
         {index.current === questions.length ? (
-          <div className="container-given-text">
-            <h1 key={animationKey} className="animated-text">
-              Felicitari ai terminat exercitiul!
-            </h1>
-            <Link to="/exercises">
-                <button className="animated-text">Continua</button>
-             </Link>
-          </div>
+          <ExerciseFinish animationKey={animationKey} />
         ) : (
-          <div className="container-given-text">
-            <h1 key={animationKey} className="animated-text">
-              {questions[index.current]}
-            </h1>
-          </div>
+          <ExerciseQuestions
+            animationKey={animationKey}
+            questions={questions}
+            index={index}
+          />
         )}
 
-{index.current === questions.length ? "" :
-<>
-        <div className="container-mid">
-          <svg
-            className="back-arrow"
-            onClick={() => (speechEnd ? handleBack() : "")}
-            xmlns="http://www.w3.org/2000/svg"
-            width="40"
-            height={index.current == 0 || !speechEnd ? "0" : "40"}
-            viewBox="0 0 175 255"
-            fill="#black"
-            transform="rotate(180)"
-          >
-            <path
-              d="M47.7037 0L0 47.7037L79.5061 127.21L0 206.716L47.7037 254.419L174.913 127.21L47.7037 0Z"
-              fill="#E5E580"
-            />
-          </svg>
+        {index.current === questions.length ? (
+          ""
+        ) : (
+          <>
+            <div className="container-mid">
+              <svg
+                className="back-arrow"
+                onClick={() => (speechEnd ? handleBack() : "")}
+                xmlns="http://www.w3.org/2000/svg"
+                width="40"
+                height={index.current == 0 || !speechEnd ? "0" : "40"}
+                viewBox="0 0 175 255"
+                fill="#black"
+                transform="rotate(180)"
+              >
+                <path
+                  d="M47.7037 0L0 47.7037L79.5061 127.21L0 206.716L47.7037 254.419L174.913 127.21L47.7037 0Z"
+                  fill="#E5E580"
+                />
+              </svg>
 
-          <p className="similar-container">
-            Similaritate: {similar.toFixed(1)}% <br />
-            {feeling}
-          </p>
-          {/* Make function to add the overall score somewhere and reset it */}
+              <p className="similar-container">
+                Similaritate: {similar.toFixed(1)}% <br />
+                {useSetFeelingValue}
+              </p>
+              {/* Make function to add the overall score somewhere and reset it */}
 
-          <svg
-            onClick={() => (speechEnd ? handleFront() : "")}
-            className="front-arrow"
-            xmlns="http://www.w3.org/2000/svg"
-            width="40"
-            height={!speechEnd ? "0" : "40"}
-            viewBox="0 0 175 255"
-            fill="#black"
-          >
-            <path
-              d="M47.7037 0L0 47.7037L79.5061 127.21L0 206.716L47.7037 254.419L174.913 127.21L47.7037 0Z"
-              fill="#E5E580"
-            />
-          </svg>
-        </div>
-        
-        <div className="container-btns">
-          {" "}
-          <button
-            style={{ display: speechEnd ? "flex" : "none" }}
-            className="start-btn"
-            onClick={handleVoiceInputCallback}
-          >
-            Porniți Înregistrarea Vocală
-          </button>
-          <button
-            style={{ display: speechEnd ? "none" : "flex" }}
-            className="stop"
-          >
-            Stop
-          </button>
-        </div>
-        </>
-        }
+              <svg
+                onClick={() => (speechEnd ? handleFront() : "")}
+                className="front-arrow"
+                xmlns="http://www.w3.org/2000/svg"
+                width="40"
+                height={!speechEnd ? "0" : "40"}
+                viewBox="0 0 175 255"
+                fill="#black"
+              >
+                <path
+                  d="M47.7037 0L0 47.7037L79.5061 127.21L0 206.716L47.7037 254.419L174.913 127.21L47.7037 0Z"
+                  fill="#E5E580"
+                />
+              </svg>
+            </div>
+
+            <div className="container-btns">
+              {" "}
+              <button
+                style={{ display: speechEnd ? "flex" : "none" }}
+                className="start-btn"
+                onClick={handleVoiceInputCallback}
+              >
+                Porniți Înregistrarea Vocală
+              </button>
+              <button
+                style={{ display: speechEnd ? "none" : "flex" }}
+                className="stop"
+              >
+                Stop
+              </button>
+            </div>
+          </>
+        )}
         {spokenText && (
           <div className="spoken-text">
             <h2 className="h2-title"> Dictie-metrul a inteles:</h2>
