+++ dictie/src/utils/manageScore.tsx
@@ -4,10 +4,9 @@ function manageScore(similar:number) {
 
     if(similar === 0) return
     else if(similar < 90){
-        console.log("NO POINT");
       
         return "no point"
-    }else console.log("POINT");
+    }else 
 
      return "point"
